Move PimcoreContextGuesser and deprecated PimcoreContextResolverAwareInterface to Pimcore\Http\Context namespace
