+++ pimcore/lib/Pimcore/Bundle/CoreBundle/Resources/config/request_response.yml
@@ -11,7 +11,7 @@ services:
     Pimcore\Http\RequestMatcherFactory:
         public: false
 
-    Pimcore\Service\Context\PimcoreContextGuesser:
+    Pimcore\Http\Context\PimcoreContextGuesser:
         public: false
