+++ pimcore/lib/Pimcore/Http/Context/PimcoreContextGuesser.php
@@ -12,7 +12,7 @@
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
-namespace Pimcore\Service\Context;
+namespace Pimcore\Http\Context;
 
 use Pimcore\Http\RequestMatcherFactory;
 use Symfony\Component\HttpFoundation\Request;
